Added "JObject" property to the "StravaObject" class
The "StravaObject" object class already inherits a "JObject" property from the "JsonObjectBase" class, but it's nullable. Since the "json" parameter of the "StravaObject" constructor isn't nullable, we know that the inherited "JObject" property is also never null. To reflect this "StravaObject.JObject" simply gets the value from the base property, but returns the value as not nullable.